+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+# 2.12.0
+
+* Refactor: Rename LibraryCollectionKey to LibraryItemKey.
+* Added LibraryContainerLocator.
+
 # 2.11.0
 
 * Added LibraryCollectionKey and LibraryCollectionLocator